+++ .idea/.name
@@ -1,1 +1,1 @@
-Overwatch Competitive Stats Tracker
+OCST
