+++ .idea/modules.xml
@@ -2,7 +2,7 @@
 <project version="4">
   <component name="ProjectModuleManager">
     <modules>
-      <module fileurl="file://$PROJECT_DIR$/OverwatchCompetitiveStatsTracker.iml" filepath="$PROJECT_DIR$/OverwatchCompetitiveStatsTracker.iml" />
+      <module fileurl="file://$PROJECT_DIR$/OCST.iml" filepath="$PROJECT_DIR$/OCST.iml" />
       <module fileurl="file://$PROJECT_DIR$/app/app.iml" filepath="$PROJECT_DIR$/app/app.iml" />
     </modules>
   </component>
