--- app/src/main/java/com/statstracker/something/ocst/UnofficialOverwatchSDK/Service/ProfileFailedEvent.java
@@ -1,7 +1,0 @@
-package com.statstracker.something.ocst.UnofficialOverwatchSDK.Service;
-
-/**
- * Created by FailQuality on 8/15/2016.
- */
-public class ProfileFailedEvent {
-}
