+++ app/src/main/java/com/statstracker/something/ocst/UnofficialOverwatchSDK/Service/ProfileService.java
@@ -2,6 +2,8 @@ package com.statstracker.something.ocst.UnofficialOverwatchSDK.Service;
 
 import com.squareup.otto.Bus;
 import com.squareup.otto.Subscribe;
+import com.statstracker.something.ocst.Events.LoadProfileCallEvent;
+import com.statstracker.something.ocst.Events.LoadProfileResponseEvent;
 import com.statstracker.something.ocst.UnofficialOverwatchSDK.API.UnofficialOverwatchAPI;
 import com.statstracker.something.ocst.UnofficialOverwatchSDK.ResponseObjects.ProfileData;
 
@@ -22,17 +24,17 @@ public class ProfileService {
     }
 
     @Subscribe
-    public void onLoadProfile(LoadProfileEvent pEvent){
+    public void onLoadProfile(LoadProfileCallEvent pEvent){
         Call<ProfileData> call = mApi.loadPlayerData(pEvent.getmPlatform(), pEvent.getmRegion(), pEvent.getmUsername());
         call.enqueue(new Callback<ProfileData>() {
             @Override
             public void onResponse(Call<ProfileData> call, Response<ProfileData> response) {
-                mBus.post(new ProfileLoadedEvent());
+                mBus.post(new LoadProfileResponseEvent(true));
             }
 
             @Override
             public void onFailure(Call<ProfileData> call, Throwable t) {
-                mBus.post(new ProfileFailedEvent());
+                mBus.post(new LoadProfileResponseEvent(false));
             }
         });
     }
